Move @anoma/wasm -> @anoma/shared
