+++ apps/namada-interface/package.json
@@ -15,10 +15,10 @@
     "@anoma/rpc": "0.1.0",
     "@anoma/seed-management": "0.1.0",
     "@anoma/session": "0.1.0",
+    "@anoma/shared": "0.1.0",
     "@anoma/tx": "0.1.0",
     "@anoma/utils": "0.1.0",
     "@anoma/wallet": "0.1.0",
-    "@anoma/wasm": "0.1.0",
     "@reduxjs/toolkit": "^1.8.0",
     "crypto-browserify": "^3.12.0",
     "framer-motion": "^6.2.8",
