+++ packages/crypto/lib/src/crypto/mod.rs
@@ -1,4 +1,3 @@
 pub mod aead;
-pub mod utils;
 pub mod mnemonic;
 pub mod bip44;
