+++ packages/crypto/lib/src/lib.rs
@@ -1,1 +1,2 @@
 pub mod crypto;
+pub mod utils;
