+++ packages/shared/.eslintrc
