+++ packages/shared/.prettierignore
