+++ packages/shared/lib/.gitignore
