+++ packages/shared/lib/.travis.yml
