+++ packages/shared/lib/LICENSE_APACHE
