+++ packages/shared/lib/LICENSE_MIT
