+++ packages/shared/lib/src/types/encrypted_keypair.rs
