+++ packages/shared/lib/src/types/mnemonic.rs
