+++ packages/shared/lib/src/wallet.rs
