+++ packages/shared/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@anoma/wasm",
+  "name": "@anoma/shared",
   "version": "0.1.0",
   "type": "module",
   "license": "MIT",
