+++ packages/tx/README.md
@@ -1,5 +1,5 @@
 # tx
 
-Anoma Wallet transaction library for interfacing with packages served from `@anoma/wasm`.
+Anoma Wallet transaction library for interfacing with packages served from `@anoma/shared`.
 
 _TBD_
