+++ packages/tx/package.json
@@ -28,7 +28,7 @@
     "typescript": "^4.7.4"
   },
   "dependencies": {
-    "@anoma/wasm": "0.1.0",
+    "@anoma/shared": "0.1.0",
     "slip44": "^1.2.60"
   }
 }
