+++ packages/wallet/package.json
@@ -16,7 +16,7 @@
   },
   "dependencies": {
     "@anoma/tx": "0.1.0",
-    "@anoma/wasm": "0.1.0",
+    "@anoma/shared": "0.1.0",
     "@cosmjs/encoding": "^0.27.1",
     "typescript": "^4.7.4"
   },
